Update for Python 3 and current-generation compilers.
Minimal, mostly-mechanical changes to make the Python programs work
correctly with Python 3.x (n.b. only tested with 3.8, which is the
version currently in Debian unstable) and to make the tests complete
cleanly with GCC 10.x and Clang 9.x.

Also update README.md to indicate that I don’t intend to do any
further work on this library myself.